+++ CHANGELOG.md
@@ -110,6 +110,7 @@
 - Add feed content type @robgietema
 - Add video content type @robgietema
 - Add push api @robgietema
+- Add ai response to search endpoint if requested @robgietema
 
 ### Bugfix